value-value scoring
dict_convert is for converting value info into dict of dict for easy access. This may need moving to the inport script.

value_scoring is my intended method for scoring values. Not properly tested because I need a full cocoa value info file. This only scores exact matches. I will make another implementation that has fuzzy scoring too using the jellyfish package. This general code can be used for the ontology code. We may want someone to cast an eye on it to see if we can speed it up.